requirements.txt
Flask-graphql
Flask
graphene==2.1.9
cassandra-driver